pyproject.toml
[tool.poetry]
name = "hass-myheat"
version = "0.2.3"
description = "MyHeat integration plugin for HomeAssistant"
authors = ["Vladimir Ermakov <vooon341@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
#python = "^3.12"
python = ">=3.12,<3.13"
#aiohttp = "^3.9"
#home-assistant-bluetooth = "^1"
#homeassistant = "^2024"
#voluptuous = ">=0.12.1"

[tool.poetry.group.dev.dependencies]
#flake8 = "*"
reorder-python-imports = "*"
#pytest = "^8"
#pytest-homeassistant-custom-component = "0.13.100"
pytest-homeassistant-custom-component = ">=0.13.100"
#bandit = "*"
black = {version = "*", allow-prereleases = true}
mypy = "*"
pre-commit = "*"
#pydocstyle = ">=6.0.0"
#prospector = {extras = ["with_vulture"], version = "*"}
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
indent = "\"    \""
not_skip = "__init__.py"
force_sort_within_sections = true
sections = "FUTURE,STDLIB,INBETWEENS,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"
known_first_party = "custom_components.myheat, tests"
combine_as_imports = true

[tool.pytest]
asyncio_mode = "auto"

[tool.pytest.ini_options]
addopts = ["-qq", "--cov=custom_components.myheat", "--asyncio-mode=auto"]
console_output_style = "count"

[tool.coverage.run]
branch = false

[tool.coverage.report]
show_missing = true
fail_under = 50

[tool.mypy]
python_version = '3.12'
ignore_errors = false
follow_imports = "silent"
ignore_missing_imports = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_configs = true
exclude='custom_components\/(?!myheat).*'

[tool.black]
line-length = 88
target-version = ['py38']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | deps
                     # the root of the project
)
| custom_components\/(?!myheat).*
'''

[tool.flake8]
exclude = ".venv,.git,.tox,docs,venv,bin,lib,deps,build"
doctests = true
# To work with Black
max-line-length = 88
# E501: line too long
# W503: Line break occurred before a binary operator
# E203: Whitespace before ':'
# D202 No blank lines allowed after function docstring
# W504 line break after binary operator
ignore = [
    "E501",
    "W503",
    "E203",
    "D202",
    "W504"
]
